requirements.txt
selenium
pytest
pytest-xdist
allure-pytest
